Try casting to uint64 for numba compat
2024-11-16T01:30:49.8365465Z hstrat/phylogenetic_inference/population/_build_distance_matrix_biopython.py:74: in build_distance_matrix_biopython
2024-11-16T01:30:49.8366452Z     distance_matrix = build_distance_matrix_numpy(
2024-11-16T01:30:49.8367213Z hstrat/phylogenetic_inference/population/_build_distance_matrix_numpy.py:68: in build_distance_matrix_numpy
2024-11-16T01:30:49.8368018Z     patristic_distance = estimate_patristic_distance_between(
2024-11-16T01:30:49.8368889Z hstrat/phylogenetic_inference/pairwise/_estimate_patristic_distance_between.py:53: in estimate_patristic_distance_between
2024-11-16T01:30:49.8369718Z     est_rank_of_mrca_between = estimate_rank_of_mrca_between(
2024-11-16T01:30:49.8371186Z hstrat/phylogenetic_inference/pairwise/_estimate_rank_of_mrca_between.py:70: in estimate_rank_of_mrca_between
2024-11-16T01:30:49.8372016Z     res = estimator(
2024-11-16T01:30:49.8372791Z hstrat/phylogenetic_inference/pairwise/_impl/_estimate_rank_of_mrca_maximum_likelihood.py:20: in estimate_rank_of_mrca_maximum_likelihood
2024-11-16T01:30:49.8373733Z     extract_common_retained_ranks_through_first_retained_disparity(
2024-11-16T01:30:49.8374848Z hstrat/phylogenetic_inference/pairwise/_impl/_extract_common_retained_ranks_through_first_retained_disparity.py:37: in extract_common_retained_ranks_through_first_retained_disparity
2024-11-16T01:30:49.8375803Z     return [
2024-11-16T01:30:49.8376583Z hstrat/juxtaposition/_impl_specimen/_iter_ranks_of_retained_commonality_between.py:13: in iter_ranks_of_retained_commonality_between
2024-11-16T01:30:49.8377538Z     for rank, comp in iter_mutual_ranks(first, second, compare=True):
2024-11-16T01:30:49.8378208Z hstrat/juxtaposition/_impl_specimen/_iter_mutual_ranks.py:62: in iter_mutual_ranks
2024-11-16T01:30:49.8378856Z     return _compare_differentia_at_common_ranks(
2024-11-16T01:30:49.8379448Z _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
2024-11-16T01:30:49.8379797Z
2024-11-16T01:30:49.8380123Z self = CPUDispatcher(<function _compare_differentia_at_common_ranks at 0x7ff98c31e0e0>)
2024-11-16T01:30:49.8381393Z args = [Array(uint8, 1, 'C', False, aligned=True), Array(uint64, 1, 'C', False, aligned=True), Array(uint8, 1, 'C', False, aligned=True), Array(uint64, 1, 'C', False, aligned=True)]
2024-11-16T01:30:49.8382340Z kws = {}
2024-11-16T01:30:49.8383125Z msg = 'No matching definition for argument type(s) array(uint8, 1d, C), array(uint64, 1d, C), array(uint8, 1d, C), array(uint64, 1d, C)'
2024-11-16T01:30:49.8383738Z
2024-11-16T01:30:49.8383923Z     def _explain_matching_error(self, *args, **kws):
2024-11-16T01:30:49.8384433Z         """
2024-11-16T01:30:49.8384868Z         Callback for the C _Dispatcher object.
2024-11-16T01:30:49.8385363Z         """
2024-11-16T01:30:49.8385777Z         assert not kws, "kwargs not handled"
2024-11-16T01:30:49.8386273Z         args = [self.typeof_pyval(a) for a in args]
2024-11-16T01:30:49.8386773Z         msg = ("No matching definition for argument type(s) %s"
2024-11-16T01:30:49.8387410Z                % ', '.join(map(str, args)))
2024-11-16T01:30:49.8387842Z >       raise TypeError(msg)
2024-11-16T01:30:49.8578034Z E       TypeError: No matching definition for argument type(s) array(uint8, 1d, C), array(uint64, 1d, C), array(uint8, 1d, C), array(uint64, 1d, C)

+++ hstrat/juxtaposition/_impl_specimen/_iter_mutual_rank_indices.py
@@ -32,10 +32,10 @@ def iter_mutual_rank_indices(
     order."""
     if compare:
         return _compare_differentia_at_common_ranks(
-            first.GetRankIndex(),
-            first.GetDifferentiaVals(),
-            second.GetRankIndex(),
-            second.GetDifferentiaVals(),
+            first.GetRankIndex().astype(np.uint64),  # numba compat
+            first.GetDifferentiaVals().astype(np.uint64),  # numba compat
+            second.GetRankIndex().astype(np.uint64),  # numba compat
+            second.GetDifferentiaVals().astype(np.uint64),  # numba compat
         )
 
     else:
